LaHuerta/backend/requirements.txt
Django==5.0.2
gunicorn==21.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
mysqlclient==2.2.0
Pillow==10.2.0 
django-sendgrid-v5==1.2.3
pytest
pytest-django
pytest-cov
